==> pipe_1_training_xgb/0_preprocess_clash.ipynb <==
cd ..

import pandas as pd
import csv

# pyensembl
from pyensembl import EnsemblRelease
import os
os.environ["PYENSEMBL_CACHE_DIR"] = "../data"
ens60 = EnsemblRelease(60)
ens60.download()
ens60.index()

# reading CLASH txt and saving as csv

with open("data/raw/clash/mmc1.txt") as f:

    # iter(f) is the same as (line for line in f)
    lines = iter(f)

    columns = []
    data = []
    for line in lines:
        if line.startswith("#"):
            line = line[1:]
            row = next(csv.reader([line], delimiter="\t"))
            columns.append(row)
        else:
            row = next(csv.reader([line], delimiter="\t"))
            data.append(row)

    # removes header text
    columns.pop(0)

    # create the output directory if it doesn"t exist
    output_dir = "."
    os.makedirs(output_dir, exist_ok=True)

    # write the output CSV files
    with open(os.path.join(output_dir, "data/raw/clash/clash_column_details.csv"), "w") as f1, \
            open(os.path.join(output_dir, "data/raw/clash/clash_raw.csv"), "w") as f2:
        writer1 = csv.writer(f1, delimiter=",")
        writer2 = csv.writer(f2, delimiter=",")
        writer1.writerows(columns)
        writer2.writerows(data)

# wrangling values

df = pd.read_csv("data/raw/clash/clash_raw.csv")

# dropping unused CLASH columns
cols_to_drop = ["miRNA_start", "miRNA_end", "chimeras_decompressed",
                "experiments", "experiments_list", "microRNA_first", "two_way_merged",
                "log2_target_enrichment", "CLASH_single_reads_ovlp",
                "5'UTR", "CDS", "3'UTR", "conservation_score",
                "CLASH_cluster_ovlp", "PAR_CLIP_cluster_ovlp"]

df.drop(columns=cols_to_drop, inplace=True)

# process microRNA_name and mRNA_name columns
new_mirna_cols = df["microRNA_name"].str.split("_", expand=True)
new_mirna_cols.columns = ["mirna_accession", "temp1", "mirna_name", "temp2"]
df = pd.concat([df, new_mirna_cols], axis=1)

new_mrna_cols = df["mRNA_name"].str.split("_", expand=True)
new_mrna_cols.columns = ["ensg", "enst", "gene_name", "temp3"]
df = pd.concat([df, new_mrna_cols], axis=1)

# dropping temporary columns
temp_cols = ["microRNA_name", "mRNA_name", "temp1", "temp2", "temp3"]
df.drop(columns=temp_cols, inplace=True)

# converting start:end biological coordinates into 0-based index
# subtracting 1 from start is enough.

df["mRNA_start"] = df["mRNA_start"] - 1
df["mRNA_end_extended"] = df["mRNA_end_extended"]

# renaming columns
rename_dict = {
    "seq_ID": "id",
    "miRNA_seq": "mirna_sequence",
    "mRNA_seq_extended": "mrna_sequence",
    "mRNA_start": "true_start_index",
    "mRNA_end_extended": "true_end_index",
    "seed_type": "true_seed_type",
    "folding_class": "true_folding_class"
}

df = df.rename(columns=rename_dict)

# augmenting df with full sequences of transcripts using pyensembl

# get sequences of ENSTs from ENSEMBL 60 to a dict
seq_dict = {
    i: ens60.transcript_by_id(i).sequence
    if ens60.transcript_by_id(i).sequence
    else None
    for i in df.enst.unique().tolist()
}

# appending full sequences
df["full_sequence_of_transcript"] = df["enst"].map(seq_dict)

# getting sequence slices from start:end positions
# if slice is closer to the 5' end and can't be extended for 30 nucleotides,
# then return the start position as 0

def get_sequence_slice(row):
    sequence = row["full_sequence_of_transcript"]
    start = max(row["true_start_index"], 0)
    end = row["true_end_index"] 
    return sequence[start:end], start, end


df[["extended_sequence", 'extended_start', "extended_end"]] = df.apply(
    lambda row: get_sequence_slice(row), axis=1, result_type='expand')


df.head()

# checking differences between fetched & df sequences

def compare_sequences(df):
    df["substring"] = df.apply(lambda row: row["full_sequence_of_transcript"][row["true_start_index"]:row["true_end_index"]], axis=1)
    df["comparison_result"] = df["substring"] == df["mrna_sequence"]
    return df

print(compare_sequences(df).comparison_result.value_counts())
print("6 rows' sequences are different")



compare_sequences(df)[compare_sequences(df)["comparison_result"] == False].gene_name.unique()

from diff_match_patch import diff_match_patch
from IPython.display import display, HTML


def highlight_differences(string1, string2):
    dmp = diff_match_patch()
    diffs = dmp.diff_main(string1, string2)
    dmp.diff_cleanupSemantic(diffs)

    highlighted_string = ""
    for op, data in diffs:
        if op == 0:  # no change
            highlighted_string += data
        elif op == -1:  # deletion
            highlighted_string += f'<del>{data}</del>'
        elif op == 1:  # insertion
            highlighted_string += f'<ins>{data}</ins>'

    return highlighted_string


# Assuming you have a DataFrame called 'df' with a 'highlighted' column
def add_color(row):
    highlighted = row['highlighted']
    if '<del>' in highlighted:
        highlighted = highlighted.replace('<del>', '<span style="color: red;">')
        highlighted = highlighted.replace('</del>', '</span>')
    if '<ins>' in highlighted:
        highlighted = highlighted.replace('<ins>', '<span style="color: green;">')
        highlighted = highlighted.replace('</ins>', '</span>')
    return highlighted

different_values_df = compare_sequences(df)[compare_sequences(df)["comparison_result"] == False]

# creating difference strings
different_values_df['highlighted'] = different_values_df.apply(lambda row: highlight_differences(row['mrna_sequence'], row['extended_sequence']), axis=1)


# render the differences
different_values_df['highlighted'] = different_values_df.apply(add_color, axis=1)

display(HTML(different_values_df[["highlighted"]].to_html(escape=False)))

print("reds symbolize sequences deleted from df sequences")
print("greens symbolize sequences added by the ENSEMBL60 sequence fetch tool")
print("whites symbolize sequences that are the same")



# extending both end of sequence for 30 nts 

def get_sequence_slice(row):
    sequence = row["full_sequence_of_transcript"]
    start = max(row["true_start_index"]-30, 0)
    end = row["true_end_index"]+30 
    return sequence[start:end], start, end


df[["extended_sequence", 'extended_start', "extended_end"]] = df.apply(
    lambda row: get_sequence_slice(row), axis=1, result_type='expand')


df.head()

# dropping temp cols and saving to csv

cols_to_drop = ["substring", "comparison_result"]

df.drop(columns=cols_to_drop, inplace=True)
df.head()

df.to_csv("data/processed/clash/clash_parsed.csv", index=False)
