==> pipe_1_training_xgb/4_merge_data.ipynb <==
%cd ..
import pandas as pd


positive_df = pd.read_csv("results/2_positive_data.csv")
positive_df["label"] = 1


negative_df = pd.read_csv("results/3_negative_data.csv")
negative_df["label"] = 0

df = pd.concat([positive_df, negative_df], ignore_index=True)
df.head()

df.isna().sum()

df.to_csv("results/4_merged_data.csv", index=False)
