==> pipe_1_training_xgb/5_generate_features.ipynb <==
%cd ..

import pandas as pd
from scripts.utils import *

pd.set_option("display.max_columns", None)

df = pd.read_csv("results/4_merged_data.csv")
df.head()

df = generate_alignment_string_from_dot_bracket(df)
df = generate_match_count_columns(df)
df = generate_ta_sps_columns(df)
df = generate_important_sites(df)
df = generate_mirna_conservation_column(df)
df = generate_seed_type_columns(df)
df = generate_mre_au_content_column(df)
df = generate_local_au_content_column(df)
df = find_mres_in_close_proximity(df)
df["position_of_mre_in_transcript"] = df["middle_of_mre"] / df["full_sequence_of_transcript"].str.len()


df.head()

df.drop(columns=["full_sequence_of_transcript"], inplace=True)
df.to_csv("results/5_generate_features.csv", index=False)
